plugins: add example low level flow logging pluging
This is an example what adding plugin examples to the Suricata repo
could look like.

More work will need to be done to make the buildable within the
tree. There is still some expectation that the headers will found in
the final installation location. But if we want to support this, we
could make it buildable in-tree.