output/null: Add the null output device
This commit adds the null output device. To prevent accidental misuse,
this feature must be configured with "--enable-nulloutput". When
configured, an output filetype of "nullsink" can be specified for the
filetype value in the outputs section.

This is implemented as an "internal eve output plugin" just like the
syslog eve output type.