eve/dns: introduce version 3 of the log
Version 2 was not reverting the IP addresses in the request as it
was using the flow direction for both message. But as the ethernet
header was reverted, this was not really making sense.

By setting version 3 of the DNS log the IP addresses are swapped
between request and answer.

Issue #6400