logging: handle directory based modules
If a module has path separators, base the module name on the path.  If
only 1 "/" (util/ioctl) log the full path as the module.  If more
(app-layer/dnp3/parser) drop the last component so a module name might
be something like "app-layer/dnp3", or "source/af-packet".

This is primarily in prep of the modular re-organization.

Additionally, actually use dn_len when building the log message to allow
for truncation of names.