detect-engine-iponly: remove insertion sort
The runtime complexity of insertion sort is approx. O(h*n)^2 where
h is the size of the HOME_NET and n is the number of ip only rules
that use the HOME_NET.

Removing this sort significantly improves rule load time when a large
HOME_NET is used in combination with a moderate amount of ip only
rules.

Brief but cursory testing shows no difference in the structure of the
radix trees that are created from this temporary linked list.