detect/cip: add check for sscanf return-value
Found by CodeQL.
Cf https://codeql.github.com/codeql-query-help/cpp/cpp-missing-check-scanf/

The sscanf function returns the number of input items successfully matched
and assigned.

We have to ensure that all subsequent uses of scanf output arguments occur in
a branch of an if statement (or similar), in which it is known that the
corresponding scanf call has in fact read all possible items from its input.
Hence, we always have to check if sscanf runs successfully by comparing
the return value to a numerical constant.

Ticket: 6753