detect/port: find unique port points
In order to create the smallest possible port ranges, it is convenient
to first have a list of unique ports. Then, the work becomes simple. See
below:

Given, a port range P1 = [1, 5]
and another, P2 = [4, 8]

right now, the code will follow a logic of recursively cutting port
ranges until we create the smallest ones. But, with the help of unique
port points, we get, unique_port_points = [1, 4, 5, 8]

So, now, we create the ranges as [1, 4), [4, 5), [5, 9) and copy the
designated SGHs where they belong. We keep these intervals right open
which means that we would not keep the high of the ranges inclusive as
otherwise they will be counted twice.