dpdk: remove redundant port checks and set copy iface socket id correctly
The function to retrieve port ID from the port name was used multiple times.
This commit removes the redundant usage of the function.

Additionally, in the DeviceConfigureIPS(), the socket ID was wrongly retrieved
for the original interface and not for the out port interface.

(cherry picked from commit 2a085ccb83bbfdd3a1ec055c952653a6713e9308)