configure.ac: get the release date from the ChangeLog
Attempt to automate the extraction and setting of the release
date. The best source we have for this is the Changelog, however that
also has its own issues.

For example, in git master, the latest ChangeLog revision is 7.0.2
which has a 2023 date. However, for release package this should always
be up to date.

The git revision date isn't ideal either.  It could be days before the
release, and the most recent commit date may not be the most recently
dated commit in the history.